Unquote paths in case they contain special characters
This happens if they e.g. contain a `:`
